Refactor, moving EncryptionUtil.
